@@ -14,11 +14,11 @@
     "@heroicons/react": "^2.0.8",
     "axios": "^0.27.2",
     "lodash": "^4.17.21",
-    "next": "12.2.5",
-    "next-auth": "^4.10.3",
+    "next": "^13.2.4",
+    "next-auth": "^4.20.1",
     "object-key-transformer": "^0.0.5",
-    "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.3",
@@ -28,7 +28,7 @@
     "@types/react-dom": "18.0.6",
     "autoprefixer": "^10.4.8",
     "eslint": "8.23.0",
-    "eslint-config-next": "12.2.5",
+    "eslint-config-next": "^13.2.4",
     "postcss": "^8.4.16",
     "tailwindcss": "^3.1.8",
     "typescript": "4.8.2"
