@@ -1,3 +1,5 @@
+"use client";
+
 import { Dialog, Disclosure, Menu, Transition } from "@headlessui/react";
 import { Bars2Icon, Bars3Icon, BellIcon, MagnifyingGlassIcon, XMarkIcon } from "@heroicons/react/20/solid";
 import Link from "next/link";
@@ -5,128 +7,7 @@ import { signIn, signOut, useSession } from "next-auth/react";
 import { Fragment, useState } from "react";
 import Image from "next/image";
 
-// const navigation = {
-//   categories: [
-//     {
-//       id: 'women',
-//       name: 'Women',
-//       featured: [
-//         {
-//           name: 'New Arrivals',
-//           href: '#',
-//           imageSrc: 'https://tailwindui.com/img/ecommerce-images/mega-menu-category-01.jpg',
-//           imageAlt: 'Models sitting back to back, wearing Basic Tee in black and bone.',
-//         },
-//         {
-//           name: 'Basic Tees',
-//           href: '#',
-//           imageSrc: 'https://tailwindui.com/img/ecommerce-images/mega-menu-category-02.jpg',
-//           imageAlt: 'Close up of Basic Tee fall bundle with off-white, ochre, olive, and black tees.',
-//         },
-//       ],
-//       sections: [
-//         {
-//           id: 'clothing',
-//           name: 'Clothing',
-//           items: [
-//             { name: 'Tops', href: '#' },
-//             { name: 'Dresses', href: '#' },
-//             { name: 'Pants', href: '#' },
-//             { name: 'Denim', href: '#' },
-//             { name: 'Sweaters', href: '#' },
-//             { name: 'T-Shirts', href: '#' },
-//             { name: 'Jackets', href: '#' },
-//             { name: 'Activewear', href: '#' },
-//             { name: 'Browse All', href: '#' },
-//           ],
-//         },
-//         {
-//           id: 'accessories',
-//           name: 'Accessories',
-//           items: [
-//             { name: 'Watches', href: '#' },
-//             { name: 'Wallets', href: '#' },
-//             { name: 'Bags', href: '#' },
-//             { name: 'Sunglasses', href: '#' },
-//             { name: 'Hats', href: '#' },
-//             { name: 'Belts', href: '#' },
-//           ],
-//         },
-//         {
-//           id: 'brands',
-//           name: 'Brands',
-//           items: [
-//             { name: 'Full Nelson', href: '#' },
-//             { name: 'My Way', href: '#' },
-//             { name: 'Re-Arranged', href: '#' },
-//             { name: 'Counterfeit', href: '#' },
-//             { name: 'Significant Other', href: '#' },
-//           ],
-//         },
-//       ],
-//     },
-//     {
-//       id: 'men',
-//       name: 'Men',
-//       featured: [
-//         {
-//           name: 'New Arrivals',
-//           href: '#',
-//           imageSrc: 'https://tailwindui.com/img/ecommerce-images/product-page-04-detail-product-shot-01.jpg',
-//           imageAlt: 'Drawstring top with elastic loop closure and textured interior padding.',
-//         },
-//         {
-//           name: 'Artwork Tees',
-//           href: '#',
-//           imageSrc: 'https://tailwindui.com/img/ecommerce-images/category-page-02-image-card-06.jpg',
-//           imageAlt:
-//             'Three shirts in gray, white, and blue arranged on table with same line drawing of hands and shapes overlapping on front of shirt.',
-//         },
-//       ],
-//       sections: [
-//         {
-//           id: 'clothing',
-//           name: 'Clothing',
-//           items: [
-//             { name: 'Tops', href: '#' },
-//             { name: 'Pants', href: '#' },
-//             { name: 'Sweaters', href: '#' },
-//             { name: 'T-Shirts', href: '#' },
-//             { name: 'Jackets', href: '#' },
-//             { name: 'Activewear', href: '#' },
-//             { name: 'Browse All', href: '#' },
-//           ],
-//         },
-//         {
-//           id: 'accessories',
-//           name: 'Accessories',
-//           items: [
-//             { name: 'Watches', href: '#' },
-//             { name: 'Wallets', href: '#' },
-//             { name: 'Bags', href: '#' },
-//             { name: 'Sunglasses', href: '#' },
-//             { name: 'Hats', href: '#' },
-//             { name: 'Belts', href: '#' },
-//           ],
-//         },
-//         {
-//           id: 'brands',
-//           name: 'Brands',
-//           items: [
-//             { name: 'Re-Arranged', href: '#' },
-//             { name: 'Counterfeit', href: '#' },
-//             { name: 'Full Nelson', href: '#' },
-//             { name: 'My Way', href: '#' },
-//           ],
-//         },
-//       ],
-//     },
-//   ],
-//   pages: [
-//     { name: 'Company', href: '#' },
-//     { name: 'Stores', href: '#' },
-//   ],
-// }
+
 const user = {
   name: 'Tom Cook',
   email: 'tom@example.com',
@@ -151,29 +32,43 @@ function classNames(...classes: String[]) {
 }
 
 const Navbar = () => {
+  // const router = useRouter();
+  // const session = useSession();
+  // const pathname = usePathname();
+
+  // if (session.status === 'authenticated') {
+  //   customAxiosInstance.setToken(session.session.accessToken as string)
+  // }
+
+  // if (pathname === '/auth/sign-in' && session.status === 'authenticated') {
+  //   router.push('/orgs')
+  // }
+
   // const [open, setOpen] = useState(false);
-  const { data, status } = useSession();
+  const { data: session, status } = useSession();
 
   const isConnected = () => status === "authenticated";
 
   return (
-    <Disclosure as="nav">
+    <Disclosure as="nav" className="bg-white">
       {({ open }) => (
         <>
           <div className="max-w-full mx-auto px-4 sm:px-6 lg:px-8">
             <div className="flex items-center justify-between h-16">
               <div className="flex items-center">
                 <div className="flex-shrink-0">
-                  <Image
-                    height={'2em'}
-                    width={'2em'}
-                    style={{
-                      maxWidth: '100%',
-                      height: 'auto',
-                    }}
-                    src="https://tailwindui.com/img/logos/workflow-mark.svg?color=indigo&shade=500"
-                    alt="Workflow"
-                  />
+                  <Link href={"/"}>
+                    <Image
+                      height={'50'}
+                      width={'50'}
+                      style={{
+                        maxWidth: '100%',
+                        height: 'auto',
+                      }}
+                      src="https://tailwindui.com/img/logos/workflow-mark.svg?color=indigo&shade=500"
+                      alt="Workflow"
+                    />
+                  </Link>
                 </div>
 
                 {/* if the user is connected */}
@@ -204,18 +99,14 @@ const Navbar = () => {
               {/* if the user is not connected */}
               {!isConnected() && (
                 <div className="hidden md:block lg:flex lg:flex-1 lg:items-center lg:justify-end lg:space-x-6">
-                  <Link href="/auth/sign-in">
-                    <a href="#" className="text-sm font-medium text-gray-700 hover:text-gray-800"onClick={() => {
+                  <Link href="/auth/sign-in" className="text-sm font-medium text-gray-700 hover:text-gray-800"onClick={() => {
                       signIn()
-                    }}> 
-                      Sign in
-                    </a>
+                  }}> 
+                    Sign in
                   </Link>
                   <span className="h-6 w-px bg-gray-200" aria-hidden="true" />
-                  <Link href="/auth/sign-up">
-                    <a href="#" className="text-sm font-medium text-gray-700 hover:text-gray-800">
-                      Create an account
-                    </a>
+                  <Link href="/auth/sign-up" className="text-sm font-medium text-gray-700 hover:text-gray-800">
+                    Create an account
                   </Link>
                 </div>
               )}
@@ -239,8 +130,8 @@ const Navbar = () => {
                         <Menu.Button className="max-w-xs bg-gray-800 rounded-full flex items-center text-sm focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-offset-gray-800 focus:ring-white">
                           <span className="sr-only">Open user menu</span>
                           <Image className="h-8 w-8 rounded-full" 
-                            height={'2em'}
-                            width={'2em'} src={user.imageUrl} alt={`${user.name} avatar`}
+                            width={'50'} height={'50'}
+                            src={user.imageUrl} alt={`${user.name} avatar`}
                           />
                         </Menu.Button>
                       </div>
@@ -373,8 +264,8 @@ const Navbar = () => {
                     <Image className="h-10 w-10 rounded-full" src={user.imageUrl} alt="" />
                   </div>
                   <div className="ml-3">
-                    <div className="text-base font-medium leading-none text-white">{data?.user?.name}</div>
-                    <div className="text-sm font-medium leading-none text-gray-400">{data?.user?.email}</div>
+                    <div className="text-base font-medium leading-none text-white">{session?.user?.name}</div>
+                    <div className="text-sm font-medium leading-none text-gray-400">{session?.user?.email}</div>
                   </div>
                   <button
                     type="button"
